core network functionalities added, interests connected to backend

+++ uitoolkit/src/main/java/com/unimapp/uitoolkit/dialogs/SimpleMultiSelectorBottomSheet.kt
@@ -79,7 +79,7 @@ class SimpleMultiSelectorBottomSheet(
     }
 
     data class Item(
-        val itemId: Int,
+        val itemId: Long,
         val itemTitle: String,
         var isSelected: Boolean
     )
